refactor: extract automations-pane.tsx into focused modules
Split the 2029-line monolith into 5 focused files:
- automations-pane.tsx (377 lines) — sidebar, list, tab routing
- automations-form-fields.tsx (182 lines) — shared form components
- automations-launch-dialog.tsx (829 lines) — launch dialog + media context
- automations-create-dialog.tsx (272 lines) — create template dialog
- automations-template-detail.tsx (424 lines) — template detail/edit pane

No behavioral changes. All exports preserved via re-exports.
