+++ .dispatch/job-state/componentizer.md
@@ -9,40 +9,38 @@ Each run of the componentizer job reads this file at Phase 0 and overwrites it a
 
 ## last_audited_sha
 
-`e59e06a60375e2a43918d7f2940e029de3c7c0e6`
+`8c264a7c53d0411edd0e922aa0deff8b762d3887`
 
 ## next_focus
 
-**File:** `apps/web/src/components/app/automations-pane.tsx` (2029 lines)
-**Strategy:** Extract subcomponents + split compound components. This file contains multiple large dialog components that are self-contained. Recommended extractions:
+**File:** `apps/web/src/components/app/feedback-panel.tsx` (1776 lines)
+**Strategy:** Extract compound components + mobile variants. This file exports 5 distinct components (ParentFeedbackPanel, FeedbackDetailPanel, ReviewSummaryPanel, MobileFeedbackSheet, MobileReviewSummarySheet). Recommended extractions:
 
-1. Extract `LaunchTemplateDialog` + `LaunchTemplateDialogContent` (~650 lines) into `automations-launch-dialog.tsx`
-2. Extract `CreateTemplateDialog` + `CreateTemplateDialogContent` (~250 lines) into `automations-create-dialog.tsx`
-3. Extract `TemplateDetail` / `TemplateDetailPane` (~360 lines) into `automations-template-detail.tsx`
+1. Extract `MobileFeedbackSheet` + `MobileReviewSummarySheet` (~430 lines) into `feedback-mobile.tsx`
+2. Extract `ReviewSummaryPanel` (~240 lines) into `review-summary-panel.tsx`
+3. Extract `useFeedbackData` hook into `use-feedback-data.ts` if a clear data-fetching hook exists
 
-This is the second-worst offender by line count and follows the same "dialogs in parent file" pattern we just addressed in jobs-pane.
+This follows the "mobile variants double component count" pattern identified in prior runs.
 
 ## backlog
 
 1. **`apps/web/src/components/app/docs-pane.tsx`** — 1880 lines. Mostly static content organized as `SECTIONS` array. Large by nature (documentation), but the section content blocks could be moved to individual files under a `docs/` directory. Lower priority since it's relatively cohesive and rarely edited for logic changes.
 
-2. **`apps/web/src/components/app/feedback-panel.tsx`** — 1776 lines. Exports 5 distinct components (ParentFeedbackPanel, FeedbackDetailPanel, ReviewSummaryPanel, MobileFeedbackSheet, MobileReviewSummarySheet). Extract mobile sheets into `feedback-mobile.tsx` (~430 lines). Extract `ReviewSummaryPanel` into `review-summary-panel.tsx` (~240 lines). Extract `useFeedbackData` hook into `use-feedback-data.ts`.
+2. **`apps/web/src/components/app/create-agent-dialog.tsx`** — 1446 lines. Large dialog with complex form state. Extract form sections into subcomponents. Extract state management into a custom hook.
 
-3. **`apps/web/src/components/app/create-agent-dialog.tsx`** — 1446 lines. Large dialog with complex form state. Extract form sections into subcomponents. Extract state management into a custom hook.
+3. **`apps/web/src/components/app/activity-pane.tsx`** — 1142 lines. Many chart subcomponents (Heatmap, ActiveHoursGrid, DailyStackedBarChart, DailyTokenChart, ModelBreakdown, ProjectBreakdown). Extract chart components into `activity-charts.tsx` (~500 lines). Extract utility functions into `activity-utils.ts`.
 
-4. **`apps/web/src/components/app/activity-pane.tsx`** — 1142 lines. Many chart subcomponents (Heatmap, ActiveHoursGrid, DailyStackedBarChart, DailyTokenChart, ModelBreakdown, ProjectBreakdown). Extract chart components into `activity-charts.tsx` (~500 lines). Extract utility functions into `activity-utils.ts`.
+4. **`apps/web/src/components/app/release-manager.tsx`** — 1142 lines. Mixed concerns: release UI, update checking, version comparison. Extract subcomponents by release lifecycle phase.
 
-5. **`apps/web/src/components/app/release-manager.tsx`** — 1142 lines. Mixed concerns: release UI, update checking, version comparison. Extract subcomponents by release lifecycle phase.
+5. **`apps/web/src/components/app/agent-history-tab.tsx`** — 1141 lines. Multiple subcomponents (AgentHistoryList, EventTimeline, FeedbackTimeline, DetailTabs, AgentHistoryDetail). Extract `AgentHistoryDetail` + `DetailTabs` into `agent-history-detail.tsx` (~340 lines). Extract timeline components into `agent-history-timeline.tsx`.
 
-6. **`apps/web/src/components/app/agent-history-tab.tsx`** — 1141 lines. Multiple subcomponents (AgentHistoryList, EventTimeline, FeedbackTimeline, DetailTabs, AgentHistoryDetail). Extract `AgentHistoryDetail` + `DetailTabs` into `agent-history-detail.tsx` (~340 lines). Extract timeline components into `agent-history-timeline.tsx`.
+6. **`apps/web/src/components/app/agents-view.tsx`** — 1026 lines. Single exported component but likely has distinct UI sections. Assess on future run.
 
-7. **`apps/web/src/components/app/agents-view.tsx`** — 1026 lines. Single exported component but likely has distinct UI sections. Assess on future run.
+7. **`apps/web/src/components/app/settings-pane.tsx`** — 840 lines. Already delegates to sub-settings components but still large. Lower priority since it's already partially extracted.
 
-8. **`apps/web/src/components/app/settings-pane.tsx`** — 840 lines. Already delegates to sub-settings components but still large. Lower priority since it's already partially extracted.
+8. **`apps/web/src/components/app/agent-card.tsx`** — 727 lines. Moderate size, single cohesive component. Lower priority.
 
-9. **`apps/web/src/components/app/agent-card.tsx`** — 727 lines. Moderate size, single cohesive component. Lower priority.
-
-10. **`apps/web/src/components/app/jobs-pane.tsx`** — 1599 lines (reduced from 2489). Further extraction possible: SettingsTab + RemoveJobDialog + PromptTab (~500 lines) into `jobs-settings-tab.tsx`, HistoryTab + RunReport (~150 lines) into `jobs-history-tab.tsx`. Defer to a later run since it just went through a split.
+9. **`apps/web/src/components/app/jobs-pane.tsx`** — 1599 lines (reduced from 2489). Further extraction possible: SettingsTab + RemoveJobDialog + PromptTab (~500 lines) into `jobs-settings-tab.tsx`, HistoryTab + RunReport (~150 lines) into `jobs-history-tab.tsx`. Defer to a later run since it just went through a split.
 
 ## patterns
 
@@ -51,9 +49,10 @@ This is the second-worst offender by line count and follows the same "dialogs in
 - **Chart/visualization code is inlined.** Recharts components with config, data transformation, and formatting helpers are kept alongside the main component instead of being pulled into dedicated chart files.
 - **Mobile variants double component count.** feedback-panel has both desktop and mobile versions of the same panels, doubling the file size.
 - **Utility functions live at the top of component files.** Formatters, comparators, and helpers (formatDate, statusClasses, shortPath) could move to shared utils or colocated `*-utils.ts` files.
-- **Shared form components get inlined.** Toggle switches, checkbox options, and other reusable form primitives get copy-pasted rather than extracted (now addressed in jobs-utils.tsx as a model).
+- **Shared form components get inlined.** Toggle switches, checkbox options, and other reusable form primitives get copy-pasted rather than extracted (now addressed in jobs-utils.tsx and automations-form-fields.tsx as models).
 
 ## history
 
 - 2026-05-17: Bootstrap scan completed. Identified 10 candidates, prioritized by impact and extraction clarity. No refactoring performed.
 - 2026-05-17: Refactored `jobs-pane.tsx` (2489→1599 lines). Extracted `jobs-helpers.ts` (pure utilities), `jobs-form-fields.tsx` (shared form components), `jobs-add-dialog.tsx` (AddJobDialog + AddJobFlow), `jobs-charts.tsx` (chart components). PR #546.
+- 2026-05-17: Refactored `automations-pane.tsx` (2029→377 lines). Extracted `automations-form-fields.tsx` (AgentTypeCombobox, TemplateWorktreeOption, TemplateFullAccessOption), `automations-launch-dialog.tsx` (LaunchTemplateDialog + context/file handling), `automations-create-dialog.tsx` (CreateTemplateDialog), `automations-template-detail.tsx` (TemplateDetailPane + TemplateDetail).
